CSV started
I set up the CSV for leaplist. A lot of the comments will need to be cleaned up and removed, but I wanted to write down things we should consider as we continue to work on this. Currently, you can add tasks and view tasks based on the current date. Let me know if you have any questions.