Add Task Button
I wasn't able to add much, but I got to mess around with tkinter and understand it much better. I added a frame within the today frame which holds the add task button. I need to finish implementing the enter task command for the button, but the goal is to create the gui for adding a task there and call the add new task function from the csv to save and update.