Added Spinbox
When the checkbox is clicked on a task it now creates a popup where the user uses spinboxes to input the hrs/mins to complete the task. The llcsv task_completed is then called and updated accordingly. We still need to figure out how to move completed tasks from the today window to the completed window. I also clean up old / redundant comments from the csv file.