Update productivity frame
I added all the dropdowns and the check box for the visualizations. It's not pretty, but it somewhat works. I think my communication/understanding of the plotCSV is causing issues and we need to refine how the two communciate. It should not return a png, but a file path instead. It definietly needs some polishing. I also commented out a fair amount of stuff that I think was causing me errors, will need to add them back in later if needed.